Rename query files to have consistent suffix.
