+++ extensions/ql-vscode/src/data-extensions-editor/external-api-usage-queries.ts
